requirements.txt
anthropic>=0.49.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-ads>=24.0.0
streamlit>=1.35.0
